# Optimize campaign articles using deferred join for pagination v2