--- a/app/views/campaigns/articles.html.haml
+++ b/app/views/campaigns/articles.html.haml
@@ -59,18 +59,23 @@
             = t("#{@presenter.course_string_prefix}.courses")
             %span.sortable-indicator
 
-      - articles = @presenter.campaign_articles
+      -  result = @presenter.campaign_articles
+      -  articles_courses = result[:articles_courses]
+      -  courses = result[:courses]
+      -  articles = result[:articles]
 
-      = will_paginate articles
+      = will_paginate articles_courses
 
       %tbody.list
-        - if articles.empty?
+        - if articles_courses.empty?
           %tr.disabled
             %td{:class => "text-center", :colSpan => 7}
               %span= I18n.t('articles.none')
         - else
-          - articles.each do |ac|
+          - articles_courses.each do |ac|
+            - course = courses[ac.course_id]
+            - article = articles[ac.article_id]
             - cache [ac, locale] do
-              = render 'campaigns/article_row', article: ac.article, article_course: ac
+              = render 'campaigns/article_row', article:, article_course: ac, course:
 
-    = will_paginate articles
+    = will_paginate articles_courses