--- a/app/views/tagged_courses/articles.html.haml
+++ b/app/views/tagged_courses/articles.html.haml
@@ -72,7 +72,8 @@
               %span= I18n.t('articles.none')
         - else
           - articles.each do |ac|
+            - course = ac.course
             - cache [ac, locale] do
-              = render 'campaigns/article_row', article: ac.article, article_course: ac
+              = render 'campaigns/article_row', article: ac.article, article_course: ac, course:
 
     = will_paginate articles