@Abishekcs 2025-07-28
**The main functionality of this PR remains the same as PR #6424, except for the addition of a test to verify that critical attributes—specifically `updated_at` from the `ArticlesCourses` table—are selected, as they are required for caching and some other minor checks which i thought will be useful**


The only updates are in the two new commits below:
- [9ba6d41](https://github.com/WikiEducationFoundation/WikiEduDashboard/pull/6430/commits/9ba6d41e63bd33b4d9f57b875b15cc81327809a2)
- [fb664cf](https://github.com/WikiEducationFoundation/WikiEduDashboard/pull/6430/commits/fb664cf9eb7c2e5f0b829e15d49379fc0b007849)


I went through the code again and tested both the feature and the test code locally. Hopefully, everything works as expected in production this time.😅