@gabina 2025-09-09
While working on PR #6460 to fix a bug, I noticed `campaigns/articles` displays many duplicate rows. I don’t think these duplicates exist in the database, and the queries don’t show any obvious cause. I couldn't reproduce the error locally, but it's easy to find in production, for both new and old campaigns:

- [data_dan_teknologi_wikimedia_indonesia_2025](https://outreachdashboard.wmflabs.org/campaigns/data_dan_teknologi_wikimedia_indonesia_2025/articles) campaign
- [africa_wiki_challenge_2022](https://outreachdashboard.wmflabs.org/campaigns/africa_wiki_challenge_2022/articles) campaign

It would make sense to see duplicate articles if they were edited for different courses in the same campaign, but here the duplicates are for the same article and the same course.

I’m not even sure if this is related to this PR though. Any ideas? I wonder if it's related to some loading/cache thing.

<img width="1201" height="1004" alt="image" src="https://github.com/user-attachments/assets/f68049fc-5c36-464c-a5b8-8d617bb3d300" />